refactor useToggle and useChange into sub modules
